requirements.txt
attrs==21.4.0
aws-cdk-lib==2.25.0
boto3==1.20.45
botocore==1.23.45
cattrs==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.10
constructs==10.1.16
DateTime==4.4
decorator==4.4.2
exceptiongroup==1.0.0rc7
idna==3.3
imageio==2.14.1
imageio-ffmpeg==0.4.5
jmespath==0.10.0
jsii==1.59.0
moviepy==1.0.3
numpy==1.21.5
opencv-python==4.5.1.48
Pillow==9.0.0
proglog==0.1.9
publication==0.0.3
pydub==0.25.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
tqdm==4.62.3
typing_extensions==4.2.0
urllib3==1.26.8
zope.interface==5.4.0
